requirements.txt
asyncio==3.4.3
azure-iot-device==2.3.0
azure-iot-hub==2.2.3
